requirements.txt
python-dotenv==1.0.1
pyserial==3.5
PyQt6==6.5.0